requirements.txt
fastapi
uvicorn
playwright
bs4
requests
google-generativeai
pyyaml